+++ backend/canisters/local_user_index/impl/src/model/local_user_map.rs
@@ -37,10 +37,6 @@ impl LocalUserMap {
     pub fn len(&self) -> usize {
         self.users.len()
     }
-
-    pub fn remove(&mut self, user_id: &UserId) {
-        self.users.remove(user_id);
-    }
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq)]