+++ frontend/openchat-agent/src/services/group/candid/idl.js
@@ -575,31 +575,46 @@ export const idlFactory = ({ IDL }) => {
     'message_index' : MessageIndex,
   });
   const PermissionRole = IDL.Variant({
+    'None' : IDL.Null,
     'Moderators' : IDL.Null,
     'Owner' : IDL.Null,
     'Admins' : IDL.Null,
     'Members' : IDL.Null,
   });
+  const CustomPermission = IDL.Record({
+    'subtype' : IDL.Text,
+    'role' : PermissionRole,
+  });
+  const MessagePermissions = IDL.Record({
+    'audio' : IDL.Opt(PermissionRole),
+    'video' : IDL.Opt(PermissionRole),
+    'custom' : IDL.Vec(CustomPermission),
+    'file' : IDL.Opt(PermissionRole),
+    'poll' : IDL.Opt(PermissionRole),
+    'text' : IDL.Opt(PermissionRole),
+    'crypto' : IDL.Opt(PermissionRole),
+    'giphy' : IDL.Opt(PermissionRole),
+    'default' : PermissionRole,
+    'image' : IDL.Opt(PermissionRole),
+    'prize' : IDL.Opt(PermissionRole),
+  });
   const GroupPermissions = IDL.Record({
-    'block_users' : PermissionRole,
     'mention_all_members' : PermissionRole,
-    'change_permissions' : PermissionRole,
     'delete_messages' : PermissionRole,
-    'send_messages' : PermissionRole,
     'remove_members' : PermissionRole,
     'update_group' : PermissionRole,
+    'message_permissions' : MessagePermissions,
     'invite_users' : PermissionRole,
+    'thread_permissions' : IDL.Opt(MessagePermissions),
     'change_roles' : PermissionRole,
     'add_members' : PermissionRole,
-    'create_polls' : PermissionRole,
     'pin_messages' : PermissionRole,
-    'reply_in_thread' : PermissionRole,
     'react_to_messages' : PermissionRole,
   });
   const PermissionsChanged = IDL.Record({
     'changed_by' : UserId,
-    'old_permissions' : GroupPermissions,
-    'new_permissions' : GroupPermissions,
+    'old_permissions_v2' : GroupPermissions,
+    'new_permissions_v2' : GroupPermissions,
   });
   const MembersAddedToDefaultChannel = IDL.Record({ 'count' : IDL.Nat32 });
   const GroupFrozen = IDL.Record({
@@ -1064,9 +1079,9 @@ export const idlFactory = ({ IDL }) => {
   });
   const GroupCanisterGroupChatSummary = IDL.Record({
     'is_public' : IDL.Bool,
-    'permissions' : GroupPermissions,
     'metrics' : ChatMetrics,
     'subtype' : IDL.Opt(GroupSubtype),
+    'permissions_v2' : GroupPermissions,
     'date_last_pinned' : IDL.Opt(TimestampMillis),
     'min_visible_event_index' : EventIndex,
     'gate' : IDL.Opt(AccessGate),
@@ -1123,9 +1138,9 @@ export const idlFactory = ({ IDL }) => {
   });
   const GroupCanisterGroupChatSummaryUpdates = IDL.Record({
     'is_public' : IDL.Opt(IDL.Bool),
-    'permissions' : IDL.Opt(GroupPermissions),
     'metrics' : IDL.Opt(ChatMetrics),
     'subtype' : GroupSubtypeUpdate,
+    'permissions_v2' : IDL.Opt(GroupPermissions),
     'date_last_pinned' : IDL.Opt(TimestampMillis),
     'gate' : AccessGateUpdate,
     'name' : IDL.Opt(IDL.Text),
@@ -1228,19 +1243,40 @@ export const idlFactory = ({ IDL }) => {
     'UserSuspended' : IDL.Null,
     'SuccessV2' : PushEventResult,
   });
+  const PermissionRoleUpdate = IDL.Variant({
+    'NoChange' : IDL.Null,
+    'SetToNone' : IDL.Null,
+    'SetToSome' : PermissionRole,
+  });
+  const OptionalMessagePermissions = IDL.Record({
+    'custom_updated' : IDL.Vec(CustomPermission),
+    'audio' : PermissionRoleUpdate,
+    'video' : PermissionRoleUpdate,
+    'file' : PermissionRoleUpdate,
+    'poll' : PermissionRoleUpdate,
+    'text' : PermissionRoleUpdate,
+    'crypto' : PermissionRoleUpdate,
+    'giphy' : PermissionRoleUpdate,
+    'custom_deleted' : IDL.Vec(IDL.Text),
+    'default' : IDL.Opt(PermissionRole),
+    'image' : PermissionRoleUpdate,
+    'prize' : PermissionRoleUpdate,
+  });
+  const OptionalMessagePermissionsUpdate = IDL.Variant({
+    'NoChange' : IDL.Null,
+    'SetToNone' : IDL.Null,
+    'SetToSome' : OptionalMessagePermissions,
+  });
   const OptionalGroupPermissions = IDL.Record({
-    'block_users' : IDL.Opt(PermissionRole),
     'mention_all_members' : IDL.Opt(PermissionRole),
-    'change_permissions' : IDL.Opt(PermissionRole),
     'delete_messages' : IDL.Opt(PermissionRole),
-    'send_messages' : IDL.Opt(PermissionRole),
     'remove_members' : IDL.Opt(PermissionRole),
     'update_group' : IDL.Opt(PermissionRole),
+    'message_permissions' : IDL.Opt(OptionalMessagePermissions),
     'invite_users' : IDL.Opt(PermissionRole),
+    'thread_permissions' : OptionalMessagePermissionsUpdate,
     'change_roles' : IDL.Opt(PermissionRole),
-    'create_polls' : IDL.Opt(PermissionRole),
     'pin_messages' : IDL.Opt(PermissionRole),
-    'reply_in_thread' : IDL.Opt(PermissionRole),
     'react_to_messages' : IDL.Opt(PermissionRole),
   });
   const UpdatedRules = IDL.Record({
@@ -1259,7 +1295,7 @@ export const idlFactory = ({ IDL }) => {
     'SetToSome' : Document,
   });
   const UpdateGroupV2Args = IDL.Record({
-    'permissions' : IDL.Opt(OptionalGroupPermissions),
+    'permissions_v2' : IDL.Opt(OptionalGroupPermissions),
     'gate' : AccessGateUpdate,
     'name' : IDL.Opt(IDL.Text),
     'description' : IDL.Opt(IDL.Text),