+++ frontend/openchat-agent/src/services/group/candid/types.d.ts
@@ -249,9 +249,9 @@ export interface CommunityCanisterChannelSummary {
   'latest_message_sender_display_name' : [] | [string],
   'channel_id' : ChannelId,
   'is_public' : boolean,
-  'permissions' : GroupPermissions,
   'metrics' : ChatMetrics,
   'subtype' : [] | [GroupSubtype],
+  'permissions_v2' : GroupPermissions,
   'date_last_pinned' : [] | [TimestampMillis],
   'min_visible_event_index' : EventIndex,
   'gate' : [] | [AccessGate],
@@ -271,9 +271,9 @@ export interface CommunityCanisterChannelSummaryUpdates {
   'latest_message_sender_display_name' : [] | [string],
   'channel_id' : ChannelId,
   'is_public' : [] | [boolean],
-  'permissions' : [] | [GroupPermissions],
   'metrics' : [] | [ChatMetrics],
   'subtype' : GroupSubtypeUpdate,
+  'permissions_v2' : [] | [GroupPermissions],
   'date_last_pinned' : [] | [TimestampMillis],
   'gate' : AccessGateUpdate,
   'name' : [] | [string],
@@ -410,6 +410,10 @@ export type Cryptocurrency = { 'InternetComputer' : null } |
 export interface CustomMessageContent {
   'data' : Uint8Array | number[],
   'kind' : string,
+}
+export interface CustomPermission {
+  'subtype' : string,
+  'role' : PermissionRole,
 }
 export type Cycles = bigint;
 export interface CyclesRegistrationFee {
@@ -642,9 +646,9 @@ export interface GovernanceProposalsSubtype {
 }
 export interface GroupCanisterGroupChatSummary {
   'is_public' : boolean,
-  'permissions' : GroupPermissions,
   'metrics' : ChatMetrics,
   'subtype' : [] | [GroupSubtype],
+  'permissions_v2' : GroupPermissions,
   'date_last_pinned' : [] | [TimestampMillis],
   'min_visible_event_index' : EventIndex,
   'gate' : [] | [AccessGate],
@@ -671,9 +675,9 @@ export interface GroupCanisterGroupChatSummary {
 }
 export interface GroupCanisterGroupChatSummaryUpdates {
   'is_public' : [] | [boolean],
-  'permissions' : [] | [GroupPermissions],
   'metrics' : [] | [ChatMetrics],
   'subtype' : GroupSubtypeUpdate,
+  'permissions_v2' : [] | [GroupPermissions],
   'date_last_pinned' : [] | [TimestampMillis],
   'gate' : AccessGateUpdate,
   'name' : [] | [string],
@@ -709,9 +713,9 @@ export interface GroupChatCreated {
 }
 export interface GroupChatSummary {
   'is_public' : boolean,
-  'permissions' : GroupPermissions,
   'metrics' : ChatMetrics,
   'subtype' : [] | [GroupSubtype],
+  'permissions_v2' : GroupPermissions,
   'date_last_pinned' : [] | [TimestampMillis],
   'min_visible_event_index' : EventIndex,
   'gate' : [] | [AccessGate],
@@ -797,19 +801,16 @@ export interface GroupNameChanged {
   'previous_name' : string,
 }
 export interface GroupPermissions {
-  'block_users' : PermissionRole,
   'mention_all_members' : PermissionRole,
-  'change_permissions' : PermissionRole,
   'delete_messages' : PermissionRole,
-  'send_messages' : PermissionRole,
   'remove_members' : PermissionRole,
   'update_group' : PermissionRole,
+  'message_permissions' : MessagePermissions,
   'invite_users' : PermissionRole,
+  'thread_permissions' : [] | [MessagePermissions],
   'change_roles' : PermissionRole,
   'add_members' : PermissionRole,
-  'create_polls' : PermissionRole,
   'pin_messages' : PermissionRole,
-  'reply_in_thread' : PermissionRole,
   'react_to_messages' : PermissionRole,
 }
 export interface GroupReactionAddedNotification {
@@ -983,6 +984,19 @@ export interface MessageMatch {
   'score' : number,
   'message_index' : MessageIndex,
 }
+export interface MessagePermissions {
+  'audio' : [] | [PermissionRole],
+  'video' : [] | [PermissionRole],
+  'custom' : Array<CustomPermission>,
+  'file' : [] | [PermissionRole],
+  'poll' : [] | [PermissionRole],
+  'text' : [] | [PermissionRole],
+  'crypto' : [] | [PermissionRole],
+  'giphy' : [] | [PermissionRole],
+  'default' : PermissionRole,
+  'image' : [] | [PermissionRole],
+  'prize' : [] | [PermissionRole],
+}
 export interface MessagePinned {
   'pinned_by' : UserId,
   'message_index' : MessageIndex,
@@ -1118,20 +1132,34 @@ export interface OptionalCommunityPermissions {
   'create_private_channel' : [] | [CommunityPermissionRole],
 }
 export interface OptionalGroupPermissions {
-  'block_users' : [] | [PermissionRole],
   'mention_all_members' : [] | [PermissionRole],
-  'change_permissions' : [] | [PermissionRole],
   'delete_messages' : [] | [PermissionRole],
-  'send_messages' : [] | [PermissionRole],
   'remove_members' : [] | [PermissionRole],
   'update_group' : [] | [PermissionRole],
+  'message_permissions' : [] | [OptionalMessagePermissions],
   'invite_users' : [] | [PermissionRole],
+  'thread_permissions' : OptionalMessagePermissionsUpdate,
   'change_roles' : [] | [PermissionRole],
-  'create_polls' : [] | [PermissionRole],
   'pin_messages' : [] | [PermissionRole],
-  'reply_in_thread' : [] | [PermissionRole],
   'react_to_messages' : [] | [PermissionRole],
 }
+export interface OptionalMessagePermissions {
+  'custom_updated' : Array<CustomPermission>,
+  'audio' : PermissionRoleUpdate,
+  'video' : PermissionRoleUpdate,
+  'file' : PermissionRoleUpdate,
+  'poll' : PermissionRoleUpdate,
+  'text' : PermissionRoleUpdate,
+  'crypto' : PermissionRoleUpdate,
+  'giphy' : PermissionRoleUpdate,
+  'custom_deleted' : Array<string>,
+  'default' : [] | [PermissionRole],
+  'image' : PermissionRoleUpdate,
+  'prize' : PermissionRoleUpdate,
+}
+export type OptionalMessagePermissionsUpdate = { 'NoChange' : null } |
+  { 'SetToNone' : null } |
+  { 'SetToSome' : OptionalMessagePermissions };
 export interface PartialUserSummary {
   'username' : [] | [string],
   'diamond_member' : boolean,
@@ -1161,14 +1189,18 @@ export interface ParticipantsRemoved {
 }
 export type PendingCryptoTransaction = { 'NNS' : NnsPendingCryptoTransaction } |
   { 'ICRC1' : Icrc1PendingCryptoTransaction };
-export type PermissionRole = { 'Moderators' : null } |
+export type PermissionRole = { 'None' : null } |
+  { 'Moderators' : null } |
   { 'Owner' : null } |
   { 'Admins' : null } |
   { 'Members' : null };
+export type PermissionRoleUpdate = { 'NoChange' : null } |
+  { 'SetToNone' : null } |
+  { 'SetToSome' : PermissionRole };
 export interface PermissionsChanged {
   'changed_by' : UserId,
-  'old_permissions' : GroupPermissions,
-  'new_permissions' : GroupPermissions,
+  'old_permissions_v2' : GroupPermissions,
+  'new_permissions_v2' : GroupPermissions,
 }
 export interface PinMessageArgs {
   'correlation_id' : bigint,
@@ -1560,7 +1592,7 @@ export type UnpinMessageResponse = { 'MessageNotFound' : null } |
   { 'UserSuspended' : null } |
   { 'SuccessV2' : PushEventResult };
 export interface UpdateGroupV2Args {
-  'permissions' : [] | [OptionalGroupPermissions],
+  'permissions_v2' : [] | [OptionalGroupPermissions],
   'gate' : AccessGateUpdate,
   'name' : [] | [string],
   'description' : [] | [string],