+++ frontend/openchat-agent/src/services/group/mappers.ts
@@ -15,6 +15,7 @@ import type {
     ApiUpdatedRules,
     ApiFollowThreadResponse,
     ApiUnfollowThreadResponse,
+    ApiOptionalMessagePermissions,
     ApiBlockUserResponse,
     ApiUnblockUserResponse,
 } from "./candid/idl";
@@ -33,7 +34,6 @@ import type {
     MemberRole,
     Message,
     GroupInviteCodeChange,
-    ChatPermissions,
     GroupCanisterGroupChatSummary,
     GroupCanisterGroupChatSummaryUpdates,
     GroupCanisterSummaryResponse,
@@ -44,6 +44,8 @@ import type {
     ConvertToCommunityResponse,
     UpdatedRules,
     FollowThreadResponse,
+    OptionalChatPermissions,
+    OptionalMessagePermissions,
 } from "openchat-shared";
 import { CommonResponses, UnsupportedValueError } from "openchat-shared";
 import type { Principal } from "@dfinity/principal";
@@ -64,7 +66,8 @@ import {
     expiredMessagesRange,
 } from "../common/chatMappers";
 import { ensureReplicaIsUpToDate } from "../common/replicaUpToDateChecker";
-import { identity, optional, optionUpdate } from "../../utils/mapping";
+import type { ApiBlockUserResponse, ApiUnblockUserResponse } from "../group/candid/idl";
+import { apiOptionUpdate, identity, optional, optionUpdate } from "../../utils/mapping";
 import { ReplicaNotUpToDateError } from "../error";
 import type { OptionalGroupPermissions } from "./candid/types";
 
@@ -118,7 +121,7 @@ function groupChatSummary(candid: ApiGroupCanisterGroupChatSummary): GroupCanist
         mentions: candid.mentions
             .filter((m) => m.thread_root_message_index.length === 0)
             .map(mention),
-        permissions: groupPermissions(candid.permissions),
+        permissions: groupPermissions(candid.permissions_v2),
         notificationsMuted: candid.notifications_muted,
         metrics: chatMetrics(candid.metrics),
         myMetrics: chatMetrics(candid.my_metrics),
@@ -166,7 +169,7 @@ function groupChatSummaryUpdates(
         mentions: candid.mentions
             .filter((m) => m.thread_root_message_index.length === 0)
             .map(mention),
-        permissions: optional(candid.permissions, groupPermissions),
+        permissions: optional(candid.permissions_v2, groupPermissions),
         notificationsMuted: optional(candid.notifications_muted, identity),
         metrics: optional(candid.metrics, chatMetrics),
         myMetrics: optional(candid.my_metrics, chatMetrics),
@@ -193,22 +196,40 @@ function updatedEvent([threadRootMessageIndex, eventIndex, timestamp]: [
 }
 
 export function apiOptionalGroupPermissions(
-    permissions: Partial<ChatPermissions>,
+    permissions: OptionalChatPermissions,
 ): OptionalGroupPermissions {
     return {
-        block_users: [],
-        change_permissions: [],
         delete_messages: apiOptional(apiPermissionRole, permissions.deleteMessages),
-        send_messages: apiOptional(apiPermissionRole, permissions.sendMessages),
         remove_members: apiOptional(apiPermissionRole, permissions.removeMembers),
         update_group: apiOptional(apiPermissionRole, permissions.updateGroup),
         invite_users: apiOptional(apiPermissionRole, permissions.inviteUsers),
         change_roles: apiOptional(apiPermissionRole, permissions.changeRoles),
-        create_polls: apiOptional(apiPermissionRole, permissions.createPolls),
         pin_messages: apiOptional(apiPermissionRole, permissions.pinMessages),
-        reply_in_thread: apiOptional(apiPermissionRole, permissions.replyInThread),
         react_to_messages: apiOptional(apiPermissionRole, permissions.reactToMessages),
         mention_all_members: apiOptional(apiPermissionRole, permissions.mentionAllMembers),
+        message_permissions: apiOptional(apiOptionalMessagePermissions, permissions.messagePermissions),
+        thread_permissions: apiOptionUpdate(apiOptionalMessagePermissions, permissions.threadPermissions)
+    };
+}
+
+function apiOptionalMessagePermissions(permissions: OptionalMessagePermissions): ApiOptionalMessagePermissions {
+    const custom_updated = permissions.memeFighter !== undefined && permissions.memeFighter !== "set_to_none"
+        ? [{ subtype: "meme_fighter", role: apiPermissionRole(permissions.memeFighter.value)}]
+        : [];
+    const custom_deleted = permissions.memeFighter === "set_to_none" ? ["meme_fighter"] : [];
+    return {
+        default: apiOptional(apiPermissionRole, permissions.default),
+        text: apiOptionUpdate(apiPermissionRole, permissions.text),
+        image: apiOptionUpdate(apiPermissionRole, permissions.image),
+        video: apiOptionUpdate(apiPermissionRole, permissions.video),
+        audio: apiOptionUpdate(apiPermissionRole, permissions.audio),
+        file: apiOptionUpdate(apiPermissionRole, permissions.file),
+        poll: apiOptionUpdate(apiPermissionRole, permissions.poll),
+        crypto: apiOptionUpdate(apiPermissionRole, permissions.crypto),
+        giphy: apiOptionUpdate(apiPermissionRole, permissions.giphy),
+        prize: apiOptionUpdate(apiPermissionRole, permissions.prize),
+        custom_updated,
+        custom_deleted,
     };
 }
 
@@ -539,8 +560,8 @@ function groupChatEvent(candid: ApiGroupChatEvent): GroupChatEvent {
     if ("PermissionsChanged" in candid) {
         return {
             kind: "permissions_changed",
-            oldPermissions: groupPermissions(candid.PermissionsChanged.old_permissions),
-            newPermissions: groupPermissions(candid.PermissionsChanged.new_permissions),
+            oldPermissions: groupPermissions(candid.PermissionsChanged.old_permissions_v2),
+            newPermissions: groupPermissions(candid.PermissionsChanged.new_permissions_v2),
             changedBy: candid.PermissionsChanged.changed_by.toString(),
         };
     }