+++ frontend/openchat-agent/src/services/online/candid/types.d.ts
@@ -195,9 +195,9 @@ export interface CommunityCanisterChannelSummary {
   'latest_message_sender_display_name' : [] | [string],
   'channel_id' : ChannelId,
   'is_public' : boolean,
-  'permissions' : GroupPermissions,
   'metrics' : ChatMetrics,
   'subtype' : [] | [GroupSubtype],
+  'permissions_v2' : GroupPermissions,
   'date_last_pinned' : [] | [TimestampMillis],
   'min_visible_event_index' : EventIndex,
   'gate' : [] | [AccessGate],
@@ -217,9 +217,9 @@ export interface CommunityCanisterChannelSummaryUpdates {
   'latest_message_sender_display_name' : [] | [string],
   'channel_id' : ChannelId,
   'is_public' : [] | [boolean],
-  'permissions' : [] | [GroupPermissions],
   'metrics' : [] | [ChatMetrics],
   'subtype' : GroupSubtypeUpdate,
+  'permissions_v2' : [] | [GroupPermissions],
   'date_last_pinned' : [] | [TimestampMillis],
   'gate' : AccessGateUpdate,
   'name' : [] | [string],
@@ -341,6 +341,10 @@ export type Cryptocurrency = { 'InternetComputer' : null } |
 export interface CustomMessageContent {
   'data' : Uint8Array | number[],
   'kind' : string,
+}
+export interface CustomPermission {
+  'subtype' : string,
+  'role' : PermissionRole,
 }
 export type Cycles = bigint;
 export interface CyclesRegistrationFee {
@@ -496,9 +500,9 @@ export interface GovernanceProposalsSubtype {
 }
 export interface GroupCanisterGroupChatSummary {
   'is_public' : boolean,
-  'permissions' : GroupPermissions,
   'metrics' : ChatMetrics,
   'subtype' : [] | [GroupSubtype],
+  'permissions_v2' : GroupPermissions,
   'date_last_pinned' : [] | [TimestampMillis],
   'min_visible_event_index' : EventIndex,
   'gate' : [] | [AccessGate],
@@ -525,9 +529,9 @@ export interface GroupCanisterGroupChatSummary {
 }
 export interface GroupCanisterGroupChatSummaryUpdates {
   'is_public' : [] | [boolean],
-  'permissions' : [] | [GroupPermissions],
   'metrics' : [] | [ChatMetrics],
   'subtype' : GroupSubtypeUpdate,
+  'permissions_v2' : [] | [GroupPermissions],
   'date_last_pinned' : [] | [TimestampMillis],
   'gate' : AccessGateUpdate,
   'name' : [] | [string],
@@ -563,9 +567,9 @@ export interface GroupChatCreated {
 }
 export interface GroupChatSummary {
   'is_public' : boolean,
-  'permissions' : GroupPermissions,
   'metrics' : ChatMetrics,
   'subtype' : [] | [GroupSubtype],
+  'permissions_v2' : GroupPermissions,
   'date_last_pinned' : [] | [TimestampMillis],
   'min_visible_event_index' : EventIndex,
   'gate' : [] | [AccessGate],
@@ -651,19 +655,16 @@ export interface GroupNameChanged {
   'previous_name' : string,
 }
 export interface GroupPermissions {
-  'block_users' : PermissionRole,
   'mention_all_members' : PermissionRole,
-  'change_permissions' : PermissionRole,
   'delete_messages' : PermissionRole,
-  'send_messages' : PermissionRole,
   'remove_members' : PermissionRole,
   'update_group' : PermissionRole,
+  'message_permissions' : MessagePermissions,
   'invite_users' : PermissionRole,
+  'thread_permissions' : [] | [MessagePermissions],
   'change_roles' : PermissionRole,
   'add_members' : PermissionRole,
-  'create_polls' : PermissionRole,
   'pin_messages' : PermissionRole,
-  'reply_in_thread' : PermissionRole,
   'react_to_messages' : PermissionRole,
 }
 export interface GroupReactionAddedNotification {
@@ -842,6 +843,19 @@ export interface MessageMatch {
   'score' : number,
   'message_index' : MessageIndex,
 }
+export interface MessagePermissions {
+  'audio' : [] | [PermissionRole],
+  'video' : [] | [PermissionRole],
+  'custom' : Array<CustomPermission>,
+  'file' : [] | [PermissionRole],
+  'poll' : [] | [PermissionRole],
+  'text' : [] | [PermissionRole],
+  'crypto' : [] | [PermissionRole],
+  'giphy' : [] | [PermissionRole],
+  'default' : PermissionRole,
+  'image' : [] | [PermissionRole],
+  'prize' : [] | [PermissionRole],
+}
 export interface MessagePinned {
   'pinned_by' : UserId,
   'message_index' : MessageIndex,
@@ -961,20 +975,34 @@ export interface OptionalCommunityPermissions {
   'create_private_channel' : [] | [CommunityPermissionRole],
 }
 export interface OptionalGroupPermissions {
-  'block_users' : [] | [PermissionRole],
   'mention_all_members' : [] | [PermissionRole],
-  'change_permissions' : [] | [PermissionRole],
   'delete_messages' : [] | [PermissionRole],
-  'send_messages' : [] | [PermissionRole],
   'remove_members' : [] | [PermissionRole],
   'update_group' : [] | [PermissionRole],
+  'message_permissions' : [] | [OptionalMessagePermissions],
   'invite_users' : [] | [PermissionRole],
+  'thread_permissions' : OptionalMessagePermissionsUpdate,
   'change_roles' : [] | [PermissionRole],
-  'create_polls' : [] | [PermissionRole],
   'pin_messages' : [] | [PermissionRole],
-  'reply_in_thread' : [] | [PermissionRole],
   'react_to_messages' : [] | [PermissionRole],
 }
+export interface OptionalMessagePermissions {
+  'custom_updated' : Array<CustomPermission>,
+  'audio' : PermissionRoleUpdate,
+  'video' : PermissionRoleUpdate,
+  'file' : PermissionRoleUpdate,
+  'poll' : PermissionRoleUpdate,
+  'text' : PermissionRoleUpdate,
+  'crypto' : PermissionRoleUpdate,
+  'giphy' : PermissionRoleUpdate,
+  'custom_deleted' : Array<string>,
+  'default' : [] | [PermissionRole],
+  'image' : PermissionRoleUpdate,
+  'prize' : PermissionRoleUpdate,
+}
+export type OptionalMessagePermissionsUpdate = { 'NoChange' : null } |
+  { 'SetToNone' : null } |
+  { 'SetToSome' : OptionalMessagePermissions };
 export interface PartialUserSummary {
   'username' : [] | [string],
   'diamond_member' : boolean,
@@ -1004,14 +1032,18 @@ export interface ParticipantsRemoved {
 }
 export type PendingCryptoTransaction = { 'NNS' : NnsPendingCryptoTransaction } |
   { 'ICRC1' : Icrc1PendingCryptoTransaction };
-export type PermissionRole = { 'Moderators' : null } |
+export type PermissionRole = { 'None' : null } |
+  { 'Moderators' : null } |
   { 'Owner' : null } |
   { 'Admins' : null } |
   { 'Members' : null };
+export type PermissionRoleUpdate = { 'NoChange' : null } |
+  { 'SetToNone' : null } |
+  { 'SetToSome' : PermissionRole };
 export interface PermissionsChanged {
   'changed_by' : UserId,
-  'old_permissions' : GroupPermissions,
-  'new_permissions' : GroupPermissions,
+  'old_permissions_v2' : GroupPermissions,
+  'new_permissions_v2' : GroupPermissions,
 }
 export type PinnedMessageUpdate = { 'NoChange' : null } |
   { 'SetToNone' : null } |