+++ backend/canisters/group_index/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.907](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.907-group_index)] - 2023-10-27
+
 ### Added
 
 - Add `permissions_v2` to `c2c_create_group` ([#4620](https://github.com/open-chat-labs/open-chat/pull/4620))