+++ backend/canisters/user_index/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.905](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.905-user_index)] - 2023-10-27
+
 ### Added
 
 - Add `permissions_v2` to group/channel summary ([#4620](https://github.com/open-chat-labs/open-chat/pull/4620))